+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/event/ClientEvents.kt
@@ -1,8 +1,11 @@
 package dev.aaronhowser.mods.geneticsresequenced.event
 
 import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
+import dev.aaronhowser.mods.geneticsresequenced.block.machines.coal_generator.CoalGeneratorMenu
 import dev.aaronhowser.mods.geneticsresequenced.control.ModKeyMappings
-import dev.aaronhowser.mods.geneticsresequenced.util.ClientUtil
+import dev.aaronhowser.mods.geneticsresequenced.packet.ModPacketHandler
+import dev.aaronhowser.mods.geneticsresequenced.packet.client_to_server.FireballPacket
+import dev.aaronhowser.mods.geneticsresequenced.packet.client_to_server.TeleportPlayerPacket
 import net.minecraft.world.inventory.AbstractContainerMenu
 import net.neoforged.api.distmarker.Dist
 import net.neoforged.bus.api.SubscribeEvent
@@ -21,11 +24,11 @@ object ClientEvents {
     @SubscribeEvent
     fun onKeyInputEvent(event: InputEvent.Key) {
         if (ModKeyMappings.TELEPORT.consumeClick()) {
-//            ModPacketHandler.messageServer(TeleportPlayerPacket())
+            ModPacketHandler.messageServer(TeleportPlayerPacket())
         }
 
         if (ModKeyMappings.DRAGONS_BREATH.consumeClick()) {
-//            ModPacketHandler.messageServer(FireballPacket())
+            ModPacketHandler.messageServer(FireballPacket())
         }
     }
 
@@ -39,7 +42,7 @@ object ClientEvents {
         val screen: AbstractContainerMenu = event.entity?.containerMenu ?: return
 
         when (screen) {
-//            is CoalGeneratorMenu -> CoalGeneratorMenu.showFuelTooltip(event)
+            is CoalGeneratorMenu -> CoalGeneratorMenu.showFuelTooltip(event)
 //            is PlasmidInfuserMenu -> PlasmidInfuserMenu.showTooltip(event)
 //            is PlasmidInjectorMenu -> PlasmidInjectorMenu.showTooltip(event)
             else -> return